+++ dskafka/integration_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/grasp-labs/ds-event-stream-go-sdk/models"
+	"github.com/grasp-labs/ds-event-stream-go-sdk/v2/models"
 	"github.com/grasp-labs/ds-go-kit/x/log"
 )